# Conflicts:
#	cooking_book/src/main/kotlin/com/example/cooking_book/controllers/Controller.kt
#	cooking_book/src/main/kotlin/com/example/cooking_book/repository/IngredientRepository.kt
#	cooking_book/src/main/kotlin/com/example/cooking_book/repository/RecipeIngredientsRepository.kt
#	cooking_book/src/main/kotlin/com/example/cooking_book/repository/RecipeRepository.kt
#	cooking_book/src/main/kotlin/com/example/cooking_book/repository/StepRepository.kt

+++ cooking_book/src/main/kotlin/com/example/cooking_book/repository/IngredientRepository.kt
@@ -5,7 +5,7 @@ import org.springframework.data.jpa.repository.JpaRepository
 import org.springframework.data.jpa.repository.Query
 import org.springframework.data.repository.query.Param
 
-interface IngredientRepository : JpaRepository<Ingredient, Long> {
+interface IngredientRepository : JpaRepository<Ingredient, Int> {
     @Query("SELECT idIngredient FROM Ingredient WHERE name = :name")
     fun getIdByName(@Param("name") name: String): Long?
 }
